+++ integration/src/main.rs
@@ -1,7 +1,8 @@
-use std::{error::Error, time::Duration};
+use std::time::Duration;
 
 use reool::{
     config::{Builder, DefaultPoolCheckoutMode},
+    error::Error,
     RedisPool,
 };
 
@@ -9,7 +10,7 @@ mod runner;
 
 #[tokio::main(threaded_scheduler)]
 #[cfg(not(feature = "basic_scheduler"))]
-async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
+async fn main() -> Result<(), Error> {
     println!("Using THREADED scheduler");
     the_real_main().await?;
     Ok(())
@@ -17,13 +18,13 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
 
 #[tokio::main(basic_scheduler)]
 #[cfg(feature = "basic_scheduler")]
-async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
+async fn main() -> Result<(), Error> {
     println!("Using BASIC scheduler");
     the_real_main().await?;
     Ok(())
 }
 
-async fn the_real_main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
+async fn the_real_main() -> Result<(), Error> {
     run_test_with_pool(1, 1).await?;
     run_test_with_pool(1, 2).await?;
     run_test_with_pool(1, 5).await?;
@@ -39,18 +40,15 @@ async fn the_real_main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
     Ok(())
 }
 
-pub async fn run_test_with_pool(
-    pool_size: usize,
-    num_pools: u32,
-) -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
-    let pool = get_builder(pool_size, num_pools).finish_redis_rs()?;
+pub async fn run_test_with_pool(pool_size: usize, num_pools: u32) -> Result<(), Error> {
+    let mut pool = get_builder(pool_size, num_pools).finish_redis_rs()?;
 
     println!(
         "=== Test with {} connection(s) in {} pool(s)",
         pool_size, num_pools,
     );
 
-    match runner::run(&pool).await {
+    match runner::run(&mut pool).await {
         Ok(()) => {
             println!("SUCCESS");
             Ok(())