+++ integration/src/runner/connections.rs
@@ -1,11 +1,8 @@
-use std::error::Error;
+use reool::{error::Error, RedisOps, RedisPool};
 
-use reool::{RedisOps, RedisPool};
-
-pub async fn run(pool: &RedisPool) -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
+pub async fn run(pool: &mut RedisPool) -> Result<(), Error> {
     println!("CONNECTIONS");
 
-    println!("Checkout 100 sequentially");
     for _ in 0u32..100 {
         let mut conn = pool.check_out_default().await?;
         conn.ping().await?;