+++ integration/src/runner/data_ops.rs
@@ -1,15 +1,11 @@
-use std::error::Error;
-
-use reool::{AsyncCommands, RedisOps, RedisPool};
+use reool::{error::Error, AsyncCommands, RedisOps, RedisPool};
 use tokio::spawn;
 
-pub async fn run(pool: &RedisPool) -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
+pub async fn run(pool: &mut RedisPool) -> Result<(), Error> {
     println!("DATA OPS");
 
-    let mut conn = pool.check_out_default().await?;
-    let db_size = conn.db_size().await?;
+    let db_size = pool.db_size().await?;
     assert_eq!(db_size, 0);
-    drop(conn);
 
     single_items_seq(pool).await?;
     single_items_par(pool).await?;
@@ -17,7 +13,7 @@ pub async fn run(pool: &RedisPool) -> Result<(), Box<dyn Error + Send + Sync + '
     Ok(())
 }
 
-async fn single_items_seq(pool: &RedisPool) -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
+async fn single_items_seq(pool: &RedisPool) -> Result<(), Error> {
     let mut conn = pool.check_out_default().await?;
 
     let v: Option<u32> = conn.get("test_key").await?;
@@ -32,87 +28,65 @@ async fn single_items_seq(pool: &RedisPool) -> Result<(), Box<dyn Error + Send +
     Ok(())
 }
 
-async fn single_items_par(pool: &RedisPool) -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
+async fn single_items_par(pool: &mut RedisPool) -> Result<(), Error> {
     let items: Vec<usize> = (0..1_000).collect();
 
     let mut handles = Vec::new();
     for i in items.iter().copied() {
         let pool = pool.clone();
         let handle = spawn(async move {
-            let mut conn = pool
-                .check_out_default()
-                .await
-                .map_err(|err| Box::new(err) as Box<dyn Error + Send + Sync + 'static>)?;
-            conn.set(i, i)
-                .await
-                .map_err(|err| Box::new(err) as Box<dyn Error + Send + Sync + 'static>)?;
-            Ok::<_, Box<dyn Error + Send + Sync + 'static>>(())
+            let mut conn = pool.check_out_default().await?;
+            conn.set(i, i).await?;
+            Ok::<_, Error>(())
         });
         handles.push(handle);
     }
 
     for handle in handles {
-        let _: () = handle.await??;
+        let _: () = handle.await.map_err(Error::cause_by)??;
     }
 
-    let mut conn = pool.check_out_default().await?;
-    let db_size = conn.db_size().await?;
-    drop(conn);
+    let db_size = pool.db_size().await?;
+
     assert_eq!(db_size, 1_000, "single_items_par set");
 
     let mut handles = Vec::new();
     for i in items.iter().copied() {
         let pool = pool.clone();
         let handle = spawn(async move {
-            let mut conn = pool
-                .check_out_default()
-                .await
-                .map_err(|err| Box::new(err) as Box<dyn Error + Send + Sync + 'static>)?;
-            let v: usize = conn
-                .get(i)
-                .await
-                .map_err(|err| Box::new(err) as Box<dyn Error + Send + Sync + 'static>)?;
+            let mut conn = pool.check_out_default().await?;
+            let v: usize = conn.get(i).await?;
             if v != i {
-                let err: Box<dyn Error + Send + Sync + 'static> = format!("{} != {}", v, i).into();
+                let err = Error::message(format!("{} != {}", v, i));
                 return Err(err);
             }
-            Ok::<_, Box<dyn Error + Send + Sync + 'static>>(())
+            Ok::<_, Error>(())
         });
         handles.push(handle);
     }
 
     for handle in handles {
-        let _: () = handle.await??;
+        let _: () = handle.await.map_err(Error::cause_by)??;
     }
 
-    let mut conn = pool.check_out_default().await?;
-    let db_size = conn.db_size().await?;
-    drop(conn);
+    let db_size = pool.db_size().await?;
     assert_eq!(db_size, 1_000, "single_items_par get");
 
     let mut handles = Vec::new();
     for i in items.iter().copied() {
-        let pool = pool.clone();
+        let mut pool = pool.clone();
         let handle = spawn(async move {
-            let mut conn = pool
-                .check_out_default()
-                .await
-                .map_err(|err| Box::new(err) as Box<dyn Error + Send + Sync + 'static>)?;
-            conn.del(i)
-                .await
-                .map_err(|err| Box::new(err) as Box<dyn Error + Send + Sync + 'static>)?;
-            Ok::<_, Box<dyn Error + Send + Sync + 'static>>(())
+            pool.del(i).await?;
+            Ok::<_, Error>(())
         });
         handles.push(handle);
     }
 
     for handle in handles {
-        let _: () = handle.await??;
+        let _: () = handle.await.map_err(Error::cause_by)??;
     }
 
-    let mut conn = pool.check_out_default().await?;
-    let db_size = conn.db_size().await?;
-    drop(conn);
+    let db_size = pool.db_size().await?;
     assert_eq!(db_size, 0, "single_items_par del");
 
     Ok(())