+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     * use `AsyncCommands` trait from `redis-rs` and remove own `Commands` trait
     * add trait `RedisOps` to support specialized use cases like ping etc.
     * reexport `redis` crate
+    * `RedisPool` implements `ConnectionLike`
 
 ## 0.25.0 (BREAKING CHANGE)
     * drop connections only on IO errors